+++ payment/service/asset/get_usdt_price.go
@@ -16,9 +16,9 @@ func (a *Asset) GetUSDTPrice(
 	return &USDTPriceModel{
 		AssetCode:          row.AssetCode,
 		AssetTitle:         row.AssetTitle,
-		Scale:              row.Scale,
+		Scale:              uint16(row.Scale),
 		ReferenceAssetCode: row.ReferenceAssetCode,
-		USDTPerAssetMinor:  row.ReferencePerAssetMinor,
+		USDTPerAssetMinor:  uint64(row.ReferencePerAssetMinor),
 		Source:             row.Source,
 		ObservedAt:         row.ObservedAt,
 		UpdatedAt:          row.UpdatedAt,
